fix: 로직 정리
